number of api calls and db users per day, week and month. naive solution
